breaking changes

+++ Sources/App/base/location/Location.swift
@@ -9,28 +9,28 @@ import Foundation
 
 public protocol Location : Hashable {
     var world : any World { get set }
-    var x : Float { get set }
-    var y : Float { get set }
-    var z : Float { get set }
+    var x : Double { get set }
+    var y : Double { get set }
+    var z : Double { get set }
     var yaw : Double { get set }
     var pitch : Double { get set }
     
-    init(world: any World, x: Float, y: Float, z: Float, yaw: Double, pitch: Double)
+    init(world: any World, x: Double, y: Double, z: Double, yaw: Double, pitch: Double)
     
     var chunk_coordinates : (x: Int, z: Int) { get }
     
     /// Whether or not the two locations are the same, regardless of `yaw` or `pitch`.
     func is_similar(_ location: Self) -> Bool
     
-    func is_nearby(center: any Location, x_radius: Float, y_radius: Float, z_radius: Float) -> Bool
+    func is_nearby(center: any Location, x_radius: Double, y_radius: Double, z_radius: Double) -> Bool
     /// Gets the distance between two locations, regardless of world.
-    func distance(to location: any Location) -> (x: Float, y: Float, z: Float)
+    func distance(to location: any Location) -> (x: Double, y: Double, z: Double)
     
     /// Returns self after adding the specified distances to self.
-    mutating func adding(x: Float, y: Float, z: Float, yaw: Double, pitch: Double) -> Self
+    mutating func adding(x: Double, y: Double, z: Double, yaw: Double, pitch: Double) -> Self
     
     /// Returns a new `Location` by adding the specified distances to this location.
-    func advanced_by(x: Float, y: Float, z: Float, yaw: Double, pitch: Double) -> Self
+    func advanced_by(x: Double, y: Double, z: Double, yaw: Double, pitch: Double) -> Self
 }
 public extension Location {
     static func == (left: any Location, right: any Location) -> Bool {
@@ -56,16 +56,16 @@ public extension Location {
         return world.uuid == location.world.uuid && x == location.x && y == location.y && z == location.z
     }
     
-    func is_nearby(center: any Location, x_radius: Float, y_radius: Float, z_radius: Float) -> Bool {
-        let (dis_x, dis_y, dis_z):(Float, Float, Float) = distance(to: center)
+    func is_nearby(center: any Location, x_radius: Double, y_radius: Double, z_radius: Double) -> Bool {
+        let (dis_x, dis_y, dis_z):(Double, Double, Double) = distance(to: center)
         return abs(dis_x) <= x_radius && abs(dis_y) <= y_radius && abs(dis_z) <= z_radius
     }
-    func distance(to location: any Location) -> (x: Float, y: Float, z: Float) {
-        let loc_x:Float = location.x, loc_y:Float = location.y, loc_z:Float = location.z
+    func distance(to location: any Location) -> (x: Double, y: Double, z: Double) {
+        let loc_x:Double = location.x, loc_y:Double = location.y, loc_z:Double = location.z
         return ((loc_x - x), (loc_y - y), (loc_z - z))
     }
     
-    mutating func adding(x: Float, y: Float, z: Float, yaw: Double, pitch: Double) -> Self {
+    mutating func adding(x: Double, y: Double, z: Double, yaw: Double, pitch: Double) -> Self {
         self.x += x
         self.y += y
         self.z += z
@@ -80,7 +80,7 @@ public extension Location {
         return self
     }
     
-    func advanced_by(x: Float, y: Float, z: Float, yaw: Double = 0, pitch: Double = 0) -> Self {
+    func advanced_by(x: Double, y: Double, z: Double, yaw: Double = 0, pitch: Double = 0) -> Self {
         return Self(world: world, x: self.x + x, y: self.y + y, z: self.z + z, yaw: self.yaw + yaw, pitch: self.pitch + pitch)
     }
 }
